+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed bug where DataLakeFileSystem.SetAccessPolicy() would throw an exception if signed identifier permissions were not in the correct order.
 - Added seekability to DataLakeFileClient.OpenRead().
 - Added additional info to exception messages.
+- Added DataLakeDirectoryClient.GetPaths().
 
 ## 12.5.0-preview.1 (2020-09-30)
 - Added support for service version 2020-02-10.